+++ lib/theoria/simp/step.ex
@@ -4,13 +4,14 @@ defmodule Theoria.Simp.Step do
   alias Theoria.Term
 
   @enforce_keys [:rule, :before, :after]
-  defstruct [:rule, :before, :after, :proof, source: :equation]
+  defstruct [:rule, :before, :after, :proof, path: [], source: :equation]
 
   @type t :: %__MODULE__{
           rule: atom(),
           before: Term.t(),
           after: Term.t(),
           proof: Term.t() | nil,
+          path: [atom()],
           source: atom()
         }
 end